FEAT: User can set custom close button on modal User can pass image, emoji, svg or text inside Modal.CloseButton component. Then passed element will be shown as close button. If nothing is passed then default close button will be shown.
