+++ react-exercises/modal-react/src/components/modal/modal-header.component.jsx
@@ -1,35 +1,7 @@
-import React, { useContext } from 'react';
-import ModalContext from './modal-context.component';
+import React from 'react';
 
 const Header = (props) => {
-  let { closeModalCallback, closeButton, backdrop, onEscapeClose } = useContext(
-    ModalContext
-  );
-
-  if (props.closeButton !== 'undefined' && typeof props.closeButton === 'boolean') {
-    closeButton = props.closeButton;
-  }
-  if (!backdrop && !onEscapeClose) {
-    closeButton = true;
-  }
-  const handleCloseButtonClick = () => {
-    closeModalCallback(false);
-  };
-  return (
-    <div className="modal-header">
-      {props.children}
-      {closeButton && (
-        <button
-          className="modal-closeBtn"
-          aria-label="modal-closeBtn"
-          data-modal-close-button="true"
-          onClick={() => handleCloseButtonClick()}
-        >
-          X
-        </button>
-      )}
-    </div>
-  );
+  return <div className="modal-header">{props.children}</div>;
 };
 
 export default Header;
