+++ react-exercises/modal-react/src/components/modal/modal.styles.css
@@ -25,6 +25,7 @@
   box-shadow: 5px 7px 20px 10px rgba(0, 0, 0, 0.2);
   overflow: auto;
   margin: auto;
+  position: relative;
 }
 
 .modal-header {
@@ -59,7 +60,7 @@
   border-bottom-left-radius: calc(0.3rem - 1px);
 }
 
-.modal-closeBtn {
+.modal-close {
   background-color: #b0b0b0;
   display: inline-block;
   -webkit-user-select: none;
@@ -72,4 +73,7 @@
   color: white;
   border-radius: 5px;
   cursor: pointer;
+  position: absolute;
+  right: 0px;
+  z-index: 999;
 }
